Rename composition helper
