+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   Add support for optional path segments.
--   Add `compose` helper for route types composition.
+-   Add `types()` helper for route types composition.
 -   Add `getUntypedParams()` method to route object.
 
 ### Fixed
