+++ src/common/createRoute/index.ts
@@ -1,1 +1,2 @@
 export * from "./createRoute.js";
+export * from "./types.js";
