--- src/common/mergeHashValues.ts
@@ -1,11 +1,0 @@
-export function mergeHashValues<T, U>(firstHash?: T[], secondHash?: U[]): (T | U)[] | undefined {
-    if (!firstHash && !secondHash) {
-        return undefined;
-    }
-
-    if (firstHash?.length === 0 || secondHash?.length === 0) {
-        return [];
-    }
-
-    return [...(firstHash ?? []), ...(secondHash ?? [])];
-}
