+++ src/common/types/hashValues.ts
