+++ src/common/types/index.ts
@@ -2,3 +2,4 @@ export * from "./type.js";
 export * from "./createType.js";
 export * from "./builtInTypes.js";
 export * from "./helpers.js";
+export * from "./hashValues.js";
